fix: consolidate web package type duplications (Phase 2)
Eliminated duplicate type definitions across web package:
- Created web-events.ts for shared event data structures (single source of truth)
- Created web-api.ts for API-specific types (Session, Agent, requests/responses)
- Created web-sse.ts for SSE-specific types (SessionEvent union with string timestamps)
- Cleaned web.ts to only contain Zod-inferred request types
- Removed 8 duplicate event data interfaces
- Deleted redundant events.ts file
- Moved lib/core.ts to types/core.ts for consistency
- Removed unnecessary re-exports from api.ts
- Updated all imports to use new structure

All 684 web package tests passing. This completes Phase 2 of the type
repair plan, significantly reducing type duplication and improving
maintainability.
