+++ packages/web/components/modals/TaskBoardModal.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react';
 import { TaskBoardModal } from './TaskBoardModal';
-import { Task } from '@/lib/core';
+import { Task } from '@/types/core';
 import { asThreadId } from '@/lib/server/core-types';
 
 // Demo columns definition for stories
