+++ packages/web/components/modals/TaskBoardModal.tsx
@@ -4,7 +4,7 @@ import React, { useState } from 'react';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faPlus, faTasks, faUser } from '@/lib/fontawesome';
 import { Modal } from '@/components/ui/Modal';
-import { Task } from '@/lib/core';
+import { Task } from '@/types/core';
 
 interface TaskBoardModalProps {
   isOpen: boolean;
