+++ packages/web/components/modals/__tests__/TaskBoardModal.test.tsx
@@ -6,7 +6,7 @@ import { render, screen } from '@testing-library/react';
 import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest';
 import '@testing-library/jest-dom/vitest';
 import { TaskBoardModal } from '@/components/modals/TaskBoardModal';
-import type { Task } from '@/lib/core';
+import type { Task } from '@/types/core';
 import { asThreadId, type ThreadId } from '@/lib/server/core-types';
 
 const mockTask: Task = {
