+++ packages/web/components/ui/AnimatedSidebar.tsx
@@ -16,8 +16,8 @@ import {
 } from '@/lib/fontawesome';
 import { ChevronDownIcon, ChevronRightIcon, ChevronLeftIcon } from '@/lib/heroicons';
 import { Timeline, RecentFile } from '@/types/design-system';
-import type { Task } from '@/lib/core';
-import type { ProjectInfo as Project } from '@/lib/core';
+import type { Task } from '@/types/core';
+import type { ProjectInfo as Project } from '@/types/core';
 import { ThemeSelector } from '@/components/ui/ThemeSelector';
 import { AccountDropdown } from '@/components/ui/AccountDropdown';
 import {
